split out utility functions
