+++ src/index.js
@@ -1,8 +1,13 @@
 const  setup = require('./setup');
-const  {match, assert} = require('./loopUI');
+const pump = require('./pump');
+const match = require('./match');
+const assert = require('./assert');
+const cgm = require('./cgm');
 
 module.exports = {
+    pump,
     match,
     assert,
-    setup
+    setup,
+    cgm
 };
